💄 조작 버튼 스타일 적용

+++ src/components/common/Sketch/index.jsx
@@ -16,7 +16,7 @@ function Sketch() {
   const [selectedShapeId, setSelectedShapeId] = useState("");
 
   useEffect(() => {
-    if (isConfirm || activeFunction === "DELETE") {
+    if (isConfirm || activeFunction === "DELETE" && shapes[0]) {
       setShapes(deleteById(shapes, selectedShapeId));
     }
   }, [isConfirm, activeFunction, shapes, selectedShapeId]);
